requirements.txt
aiofile==3.7.2
aiohttp>=3.7.4
azure-ai-formrecognizer==3.1.2
azure-identity==1.7.0
azure-storage-blob==12.9.0
coveralls==3.2.0
Django==3.2.8
django-coverage-plugin==2.0.1
django-redis==5.0.0
django-storages==1.12.2
faker==9.5.2
firebase-admin==5.0.3
gcloud-aio-storage==6.2.0
gcloud-rest-storage==6.2.0
google-api-python-client==2.27.0
google-cloud-secret-manager==2.7.2
google-cloud-storage==1.42.3
google-python-cloud-debugger==2.18
gunicorn==20.1.0
hiredis==2.0.0
lz4==3.1.3
pandas==1.3.4
Pillow==8.4.0
python-dotenv==0.19.1
pytz==2021.3
sendgrid==6.8.3
sentry-sdk==1.4.3
uvicorn==0.15.0
